compare edit song + split component to basesettings + global component
